+++ docs/examples/dropdown/doOrder.js
@@ -1,0 +1,67 @@
+// @flow strict
+import { Fragment, type Node, useState, useRef } from 'react';
+import { IconButton, Dropdown, Flex, Box, FixedZIndex, CompositeZIndex } from 'gestalt';
+
+export default function CustomIconButtonPopoverExample(): Node {
+  const PAGE_HEADER_ZINDEX = new FixedZIndex(10);
+
+  const [open, setOpen] = useState(false);
+  const [selected, setSelected] = useState(null);
+  const anchorRef = useRef(null);
+  const onSelect: $ElementType<React$ElementConfig<typeof Dropdown.Item>, 'onSelect'> = ({
+    item,
+  }) => setSelected(item);
+
+  return (
+    <Fragment>
+      <Flex justifyContent="center" width="100%" height="100%">
+        <Box margin={2}>
+          <IconButton
+            accessibilityControls="selectlist-dropdown-example3"
+            accessibilityExpanded={open}
+            accessibilityHaspopup
+            accessibilityLabel="Menu"
+            bgColor="lightGray"
+            icon="ellipsis"
+            iconColor="darkGray"
+            onClick={() => setOpen((prevVal) => !prevVal)}
+            ref={anchorRef}
+            selected={open}
+            size="lg"
+          />
+        </Box>
+      </Flex>
+      {open && (
+        <Dropdown
+          anchor={anchorRef.current}
+          id="selectlist-dropdown-example3"
+          onDismiss={() => setOpen(false)}
+          zIndex={new CompositeZIndex([PAGE_HEADER_ZINDEX])}
+        >
+          <Dropdown.Item
+            onSelect={onSelect}
+            option={{ value: 'Download image', label: 'Download image' }}
+            selected={selected}
+          />
+          <Dropdown.Item
+            badge={{ text: 'New' }}
+            onSelect={onSelect}
+            option={{ value: 'Hide Pin', label: 'Hide Pin' }}
+            selected={selected}
+          />
+          <Dropdown.Link
+            href="https://pinterest.com"
+            isExternal
+            option={{ value: 'Report Pin', label: 'Report Pin' }}
+            onClick={({ event }) => event.preventDefault()}
+          />
+          <Dropdown.Item
+            onSelect={onSelect}
+            option={{ value: 'Delete Pin', label: 'Delete Pin' }}
+            selected={selected}
+          />
+        </Dropdown>
+      )}
+    </Fragment>
+  );
+}