+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-list/invalid/no-gestalt-list-ol.js
@@ -1,0 +1,5 @@
+export default function TestElement() {
+  return <ol>
+    <li>Gestalt</li>
+  </ol>;
+}