+++ packages/gestalt/src/Dropdown.js
@@ -8,7 +8,10 @@ import { ESCAPE, SPACE, TAB, ENTER, UP_ARROW, DOWN_ARROW } from './keyCodes.js';
 import Layer from './Layer.js';
 import Popover from './Popover.js';
 import { type Indexable } from './zIndex.js';
+import AnimationProvider from './animation/AnimationContext.js';
+import { useDeviceType } from './contexts/DeviceTypeProvider.js';
 import { DropdownContextProvider } from './Dropdown/Context.js';
+import PartialPage from './SheetMobile/PartialPage.js';
 import { type DirectionOptionType } from './utils/keyboardNavigation.js';
 
 const KEYS = {
@@ -123,9 +126,17 @@ type Props = {|
    */
   maxHeight?: '30vh',
   /**
+   * Mobile-only prop. Callback fired when Dropdown's in & out animations end. See the [mobile variant](https://gestalt.pinterest.systems/web/dropdown#mobile) to learn more.
+   */
+  mobileOnAnimationEnd?: ({| animationState: 'in' | 'out' |}) => void,
+  /**
    * Callback fired when the menu is closed.
    */
   onDismiss: () => void,
+  /**
+   * Dropdown can adapt to mobile devices to [SheetMobile](https://gestalt.pinterest.systems/web/sheetmobile). Mobile adaptation is disabled by default. Set to 'false' to enable SheetMobile in mobile devices. See the [mobile variant](https://gestalt.pinterest.systems/web/dropdown#mobile) to learn more.
+   */
+  disableMobileUI?: boolean,
   /**
    * An object representing the zIndex value of the Dropdown menu. Learn more about [zIndex classes](https://gestalt.pinterest.systems/web/zindex_classes)
    */
@@ -149,8 +160,13 @@ export default function Dropdown({
   onDismiss,
   zIndex,
   maxHeight,
+  mobileOnAnimationEnd,
+  disableMobileUI = true,
 }: Props): Node {
-  const [hoveredItem, setHoveredItem] = useState<number>(0);
+  const deviceType = useDeviceType();
+  const isMobile = deviceType === 'mobile';
+
+  const [hoveredItemIndex, setHoveredItemIndex] = useState<?number>(isMobile ? undefined : 0);
 
   const dropdownChildrenArray = Children.toArray(children);
   const allowedChildrenOptions = getChildrenOptions(dropdownChildrenArray);
@@ -170,7 +186,7 @@ export default function Dropdown({
     event: SyntheticKeyboardEvent<HTMLElement>,
     direction: DirectionOptionType,
   ) => {
-    const newIndex = direction + hoveredItem;
+    const newIndex = direction + (hoveredItemIndex ?? 0);
     const optionsCount = allowedChildrenOptions.length - 1;
 
     // If there's an existing item, navigate from that position
@@ -189,7 +205,7 @@ export default function Dropdown({
 
     if (cursorOption) {
       const item = cursorOption.option;
-      setHoveredItem(cursorIndex);
+      setHoveredItemIndex(cursorIndex);
 
       if (direction === KEYS.ENTER) {
         cursorOption.onSelect?.({
@@ -218,6 +234,29 @@ export default function Dropdown({
       event.preventDefault();
     }
   };
+  if (isMobile && !disableMobileUI) {
+    return (
+      <AnimationProvider>
+        <PartialPage
+          align="start"
+          padding="default"
+          onDismiss={onDismiss}
+          onAnimationEnd={mobileOnAnimationEnd}
+          role="dialog"
+          showDismissButton
+          size="auto"
+          zIndex={zIndex}
+        >
+          {headerContent}
+          <DropdownContextProvider
+            value={{ id, hoveredItemIndex, setHoveredItemIndex, setOptionRef }}
+          >
+            {renderChildrenWithIndex(dropdownChildrenArray)}
+          </DropdownContextProvider>
+        </PartialPage>
+      </AnimationProvider>
+    );
+  }
 
   const dropdown = (
     <Popover
@@ -243,7 +282,9 @@ export default function Dropdown({
       >
         {Boolean(headerContent) && <Box padding={2}>{headerContent}</Box>}
 
-        <DropdownContextProvider value={{ id, hoveredItem, setHoveredItem, setOptionRef }}>
+        <DropdownContextProvider
+          value={{ id, hoveredItemIndex, setHoveredItemIndex, setOptionRef }}
+        >
           {renderChildrenWithIndex(dropdownChildrenArray)}
         </DropdownContextProvider>
       </Box>